LayoutTests/fast/text/font-selection-font-face-parse-expected.txt

PASS weightTestSheet.cssRules[0].style.fontWeight is "1"
PASS weightTestSheet.cssRules[1].style.fontWeight is "100"
PASS weightTestSheet.cssRules[2].style.fontWeight is "200"
PASS weightTestSheet.cssRules[3].style.fontWeight is "300"
PASS weightTestSheet.cssRules[4].style.fontWeight is "400"
PASS weightTestSheet.cssRules[5].style.fontWeight is "normal"
PASS weightTestSheet.cssRules[6].style.fontWeight is "500"
PASS weightTestSheet.cssRules[7].style.fontWeight is "600"
PASS weightTestSheet.cssRules[8].style.fontWeight is "700"
PASS weightTestSheet.cssRules[9].style.fontWeight is "bold"
PASS weightTestSheet.cssRules[10].style.fontWeight is "800"
PASS weightTestSheet.cssRules[11].style.fontWeight is "900"
PASS weightTestSheet.cssRules[12].style.fontWeight is ""
PASS weightTestSheet.cssRules[13].style.fontWeight is ""
PASS weightTestSheet.cssRules[14].style.fontWeight is ""
PASS weightTestSheet.cssRules[15].style.fontWeight is "7"
PASS weightTestSheet.cssRules[16].style.fontWeight is "300"
PASS weightTestSheet.cssRules[17].style.fontWeight is "200"
PASS stretchTestSheet.cssRules[0].style.fontStretch is "1%"
PASS stretchTestSheet.cssRules[1].style.fontStretch is "2"
PASS stretchTestSheet.cssRules[2].style.fontStretch is "ultra-condensed"
PASS stretchTestSheet.cssRules[3].style.fontStretch is "extra-condensed"
PASS stretchTestSheet.cssRules[4].style.fontStretch is "condensed"
PASS stretchTestSheet.cssRules[5].style.fontStretch is "semi-condensed"
PASS stretchTestSheet.cssRules[6].style.fontStretch is "normal"
PASS stretchTestSheet.cssRules[7].style.fontStretch is "semi-expanded"
PASS stretchTestSheet.cssRules[8].style.fontStretch is "expanded"
PASS stretchTestSheet.cssRules[9].style.fontStretch is "extra-expanded"
PASS stretchTestSheet.cssRules[10].style.fontStretch is "ultra-expanded"
PASS stretchTestSheet.cssRules[11].style.fontStretch is ""
PASS stretchTestSheet.cssRules[12].style.fontStretch is ""
PASS stretchTestSheet.cssRules[13].style.fontStretch is "7"
PASS styleTestSheet.cssRules[0].style.fontStyle is "1deg"
PASS styleTestSheet.cssRules[1].style.fontStyle is "200grad"
PASS styleTestSheet.cssRules[2].style.fontStyle is "6.28318rad"
PASS styleTestSheet.cssRules[3].style.fontStyle is "4turn"
PASS styleTestSheet.cssRules[4].style.fontStyle is "5"
PASS styleTestSheet.cssRules[5].style.fontStyle is "20"
PASS styleTestSheet.cssRules[6].style.fontStyle is "italic"
PASS styleTestSheet.cssRules[7].style.fontStyle is "oblique"
PASS styleTestSheet.cssRules[8].style.fontStyle is "normal"
PASS styleTestSheet.cssRules[9].style.fontStyle is ""
PASS styleTestSheet.cssRules[10].style.fontStyle is ""
PASS styleTestSheet.cssRules[11].style.fontStyle is "7"
PASS styleTestSheet.cssRules[12].style.fontStyle is "calc(1441deg)"
PASS successfullyParsed is true

TEST COMPLETE
